feat: 中優先度残件4件実装（protocolsend/recv, 委譲問題, Int64, modem制御）
1. protocolsend / protocolrecv 実装
   - MacroRunner に cmdProtocolSend/cmdProtocolRecv を追加
   - resolveProtocolArg でプロトコル名をバリデーション
   - 内部で cmdFileTransferSend/cmdFileTransferRecv に委譲

2. FileTransferDelegate 二重委譲問題を解決
   - XPC転送開始時にデリゲートを MacroXPCManager に切り替え
   - 転送完了/失敗時に restoreTerminalDelegate() で復帰
   - transferDidRequestSend を connectionManager.send() に転送

3. XPC 転送バイト型を Int64 に変更
   - getTransferStatus シグネチャ (String, Int, Int) → (String, Int64, Int64)
   - プロトコル・実装・モック・テスト全箇所を更新
   - 2GB超ファイル転送のオーバーフロー問題を解消

4. getModemStatus / DTR / RTS 実装
   - SerialConnection に getModemStatus() 追加 (ioctl TIOCMGET)
   - SerialConnection に setDtr()/setRts() 追加 (ioctl TIOCMBIS/TIOCMBIC)
   - MacroXPCManager.getModemStatus をシリアル接続経由で実装
   - TerminalWindowController.ttlSetDtr/ttlSetRts を実結合

テスト追加: ProtocolSendRecvTests (3), Int64TransferStatusTests (2),
ProtocolNameResolutionTests (3)

+++ TeraTermMac/Sources/TTLMacroShared/MacroXPCProtocols.swift
@@ -242,8 +242,8 @@ public enum TransferStatusString: String {
     func startFileRecv(protocolName: String, localDir: String,
                        reply: @escaping (Bool, String, String) -> Void)
 
-    /// Get transfer status
-    func getTransferStatus(reply: @escaping (String, Int, Int) -> Void)
+    /// Get transfer status (status, bytesTransferred, totalBytes)
+    func getTransferStatus(reply: @escaping (String, Int64, Int64) -> Void)
 
     /// Cancel current transfer
     func cancelTransfer(reply: @escaping () -> Void)
